+++ src/RecursiveGrowthMain.java
@@ -6,14 +6,14 @@ public class RecursiveGrowthMain {
         System.out.println("Enter the initial number of organisms: ");
         int initOrganisms = keyboard.nextInt();
         System.out.println("Enter the growth rate: ");
-        int growthRate = keyboard.nextInt();
+        double growthRate = keyboard.nextInt();
         System.out.println("Enter the time (hours) needed to achieve the growth rate: ");
-        int time = keyboard.nextInt();
+        double time = keyboard.nextInt();
         System.out.println("Enter the observation time (hours): ");
-        int obsTime = keyboard.nextInt();
-        int endOrganismsNum = initOrganisms;
+        double obsTime = keyboard.nextInt();
+        double endOrganismsNum = initOrganisms;
         for (int i = 0; i < (obsTime / time); i++)
             endOrganismsNum *= growthRate;
-        System.out.println("After " + obsTime + " hours there will be " + endOrganismsNum + " organisms.");
+        System.out.println("After " + (int)obsTime + " hours there will be " + (int)endOrganismsNum + " organisms.");
     }
 }
